@@ -4,33 +4,33 @@ import (
 	"errors"
 )
 
-type sliceList struct {
-	data []interface{}
+type sliceList[T comparable] struct {
+	data []T
 }
 
-func NewList() List {
-	return &sliceList{}
+func NewList[T comparable]() List[T] {
+	return &sliceList[T]{}
 }
 
-func (s *sliceList) Clear() {
+func (s *sliceList[T]) Clear() {
 	s.data = nil
 }
 
-func (s *sliceList) IsEmpty() bool {
+func (s *sliceList[T]) IsEmpty() bool {
 	return len(s.data) == 0
 }
 
-func (s *sliceList) Size() int {
+func (s *sliceList[T]) Size() int {
 	return len(s.data)
 }
 
-func (s *sliceList) ForEach(f func(interface{})) {
+func (s *sliceList[T]) ForEach(f func(T)) {
 	for _, v := range s.data {
 		f(v)
 	}
 }
 
-func (s *sliceList) ForEachFailFast(f func(interface{}) error) error {
+func (s *sliceList[T]) ForEachFailFast(f func(T) error) error {
 	for _, v := range s.data {
 		err := f(v)
 		if err != nil {
@@ -40,16 +40,16 @@ func (s *sliceList) ForEachFailFast(f func(interface{}) error) error {
 	return nil
 }
 
-func (s *sliceList) ForEachAsync(f func(interface{})) {
+func (s *sliceList[T]) ForEachAsync(f func(T)) {
 	s.ForEach(f)
 }
 
-func (s *sliceList) Add(v interface{}) bool {
+func (s *sliceList[T]) Add(v T) bool {
 	s.data = append(s.data, v)
 	return true
 }
 
-func (s *sliceList) AddIndex(i int, v interface{}) {
+func (s *sliceList[T]) AddIndex(i int, v T) {
 	l := len(s.data)
 
 	if i < 0 || i >= l {
@@ -57,7 +57,7 @@ func (s *sliceList) AddIndex(i int, v interface{}) {
 	}
 
 	if i == 0 {
-		s.data = append([]interface{}{v}, s.data...)
+		s.data = append([]T{v}, s.data...)
 	} else if i == l {
 		s.data = append(s.data, v)
 	} else {
@@ -66,7 +66,7 @@ func (s *sliceList) AddIndex(i int, v interface{}) {
 	}
 }
 
-func (s *sliceList) Contains(v interface{}) bool {
+func (s *sliceList[T]) Contains(v T) bool {
 	for _, e := range s.data {
 		if e == v {
 			return true
@@ -75,14 +75,14 @@ func (s *sliceList) Contains(v interface{}) bool {
 	return false
 }
 
-func (s *sliceList) Get(i int) interface{} {
+func (s *sliceList[T]) Get(i int) T {
 	if i < 0 || i >= len(s.data) {
 		panic(errors.New("index out of bounds"))
 	}
 	return s.data[i]
 }
 
-func (s *sliceList) IndexOf(v interface{}) int {
+func (s *sliceList[T]) IndexOf(v T) int {
 	for i, e := range s.data {
 		if e == v {
 			return i
@@ -91,7 +91,7 @@ func (s *sliceList) IndexOf(v interface{}) int {
 	return -1
 }
 
-func (s *sliceList) FindIndexOf(f Predicate) int {
+func (s *sliceList[T]) FindIndexOf(f Predicate[T]) int {
 	for i, e := range s.data {
 		if f(e) {
 			return i
@@ -100,7 +100,7 @@ func (s *sliceList) FindIndexOf(f Predicate) int {
 	return -1
 }
 
-func (s *sliceList) Remove(v interface{}) bool {
+func (s *sliceList[T]) Remove(v T) bool {
 	for i, e := range s.data {
 		if e == v {
 			return s.RemoveIndex(i)
@@ -109,7 +109,7 @@ func (s *sliceList) Remove(v interface{}) bool {
 	return false
 }
 
-func (s *sliceList) RemoveIndex(i int) bool {
+func (s *sliceList[T]) RemoveIndex(i int) bool {
 	if i < 0 || i >= len(s.data) {
 		return false
 	}
@@ -129,29 +129,29 @@ func (s *sliceList) RemoveIndex(i int) bool {
 	return true
 }
 
-func copySlice(s []interface{}) []interface{} {
-	var a []interface{}
+func copySlice[T any](s []T) []T {
+	var a []T
 	if s != nil {
-		a = make([]interface{}, len(s))
+		a = make([]T, len(s))
 		copy(a, s)
 	}
 	return a
 }
 
-func reverseSlice(s []interface{}) []interface{} {
+func reverseSlice[T any](s []T) []T {
 	for i, j := 0, len(s)-1; i < j; i, j = i+1, j-1 {
 		s[i], s[j] = s[j], s[i]
 	}
 	return s
 }
 
-func (s *sliceList) Iterator() Iterator {
+func (s *sliceList[T]) Iterator() Iterator[T] {
 	a := copySlice(s.data)
-	return NewIterator(a...)
+	return NewIterator[T](a...)
 }
 
-func (s *sliceList) ReverseIterator() Iterator {
+func (s *sliceList[T]) ReverseIterator() Iterator[T] {
 	a := copySlice(s.data)
 	a = reverseSlice(a)
-	return NewIterator(a...)
+	return NewIterator[T](a...)
 }
