@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 )
 
-// Get retrieves the value for a key in the bucket.
+// GetJSON retrieves the value for a key in the bucket.
 // Returns a nil value if the key does not exist or if the key is a nested bucket.
 // The returned value is only valid for the life of the transaction.
 func (b *Bucket) GetJSON(key string, val interface{}) bool {
@@ -17,7 +17,7 @@ func (b *Bucket) GetJSON(key string, val interface{}) bool {
 	return err == nil
 }
 
-// Put sets the value for a key in the bucket.
+// PutJSON sets the value for a key in the bucket.
 // If the key exist then its previous value will be overwritten.
 // Supplied value must remain valid for the life of the transaction.
 // Returns an error if the bucket was created from a read-only transaction, if the key is blank, if the key is too large, or if the value is too large.
