@@ -2,25 +2,25 @@ package util
 
 import "sync"
 
-type PriorityQueue struct {
+type PriorityQueue[T any] struct {
 	mutex   sync.Mutex
-	entries []PriorityEntry
+	entries []PriorityEntry[T]
 }
 
-type PriorityEntry struct {
+type PriorityEntry[T any] struct {
 	Priority int
-	Element  interface{}
+	Element  T
 }
 
-func (p *PriorityQueue) Add(e interface{}) {
+func (p *PriorityQueue[T]) Add(e T) {
 	p.AddPriority(0, e)
 }
 
-func (p *PriorityQueue) AddPriority(priority int, e interface{}) {
+func (p *PriorityQueue[T]) AddPriority(priority int, e T) {
 	p.mutex.Lock()
 	defer p.mutex.Unlock()
 
-	ent := PriorityEntry{Priority: priority, Element: e}
+	ent := PriorityEntry[T]{Priority: priority, Element: e}
 
 	for i, existing := range p.entries {
 		if existing.Priority > priority {
@@ -34,7 +34,7 @@ func (p *PriorityQueue) AddPriority(priority int, e interface{}) {
 	p.entries = append(p.entries, ent)
 }
 
-func (p *PriorityQueue) IsEmpty() bool {
+func (p *PriorityQueue[T]) IsEmpty() bool {
 	p.mutex.Lock()
 	defer p.mutex.Unlock()
 
@@ -42,11 +42,11 @@ func (p *PriorityQueue) IsEmpty() bool {
 }
 
 // clone returns a clone of the entry list. Used to take a snapshot, see ForEach
-func (p *PriorityQueue) clone() []PriorityEntry {
+func (p *PriorityQueue[T]) clone() []PriorityEntry[T] {
 	p.mutex.Lock()
 	defer p.mutex.Unlock()
 
-	e := make([]PriorityEntry, len(p.entries))
+	e := make([]PriorityEntry[T], len(p.entries))
 	copy(e, p.entries)
 
 	return e
@@ -54,7 +54,7 @@ func (p *PriorityQueue) clone() []PriorityEntry {
 
 // ForEach will call a function for each entry in the PriorityQueue.
 // The Queue is not modified during this call.
-func (p *PriorityQueue) ForEach(f func(interface{}) error) error {
+func (p *PriorityQueue[T]) ForEach(f func(T) error) error {
 	// Run on a clone() so the underlying queue can change whilst we are running
 	for _, e := range p.clone() {
 		err := f(e.Element)
@@ -67,7 +67,7 @@ func (p *PriorityQueue) ForEach(f func(interface{}) error) error {
 
 // Drain will call a function for each entry in the PriorityQueue,
 // removing the entry from the head of the queue.
-func (p *PriorityQueue) Drain(f func(interface{}) error) error {
+func (p *PriorityQueue[T]) Drain(f func(T) error) error {
 	for {
 		e, exists := p.Pop()
 		if !exists {
@@ -82,13 +82,15 @@ func (p *PriorityQueue) Drain(f func(interface{}) error) error {
 }
 
 // Pop removes the first entry from the queue.
-func (p *PriorityQueue) Pop() (interface{}, bool) {
+func (p *PriorityQueue[T]) Pop() (T, bool) {
 	p.mutex.Lock()
 	defer p.mutex.Unlock()
 
 	// Don't use p.IsEmpty() to save a round of locking
 	if len(p.entries) == 0 {
-		return nil, false
+		// Before generics, we could return nil, but this is the only way with generics
+		var emptyVal T
+		return emptyVal, false
 	}
 
 	e := p.entries[0]
