@@ -3,29 +3,29 @@ package util
 import "sync"
 
 // syncSet is a synchronized set with accessors similar to the Java Set interface
-type syncSet struct {
+type syncSet[T comparable] struct {
 	mutex sync.Mutex
-	m     map[interface{}]interface{}
+	m     map[T]interface{}
 }
 
 // NewSyncSet creates a new Synchronous Set
-func NewSyncSet() Set {
-	s := &syncSet{}
+func NewSyncSet[T comparable]() Set[T] {
+	s := &syncSet[T]{}
 	s.Clear()
 	return s
 }
 
-func (m *syncSet) Size() int {
+func (m *syncSet[T]) Size() int {
 	m.mutex.Lock()
 	defer m.mutex.Unlock()
 	return len(m.m)
 }
 
-func (m *syncSet) IsEmpty() bool {
+func (m *syncSet[T]) IsEmpty() bool {
 	return m.Size() == 0
 }
 
-func (m *syncSet) Add(v interface{}) bool {
+func (m *syncSet[T]) Add(v T) bool {
 	m.mutex.Lock()
 	defer m.mutex.Unlock()
 	_, e := m.m[v]
@@ -35,7 +35,7 @@ func (m *syncSet) Add(v interface{}) bool {
 	return !e
 }
 
-func (m *syncSet) AddAll(v ...interface{}) bool {
+func (m *syncSet[T]) AddAll(v ...T) bool {
 	var r bool
 	for _, e := range v {
 		r = m.Add(e) || r
@@ -43,14 +43,14 @@ func (m *syncSet) AddAll(v ...interface{}) bool {
 	return r
 }
 
-func (m *syncSet) Clear() {
+func (m *syncSet[T]) Clear() {
 	m.mutex.Lock()
 	defer m.mutex.Unlock()
 
-	m.m = make(map[interface{}]interface{})
+	m.m = make(map[T]interface{})
 }
 
-func (m *syncSet) Remove(k interface{}) bool {
+func (m *syncSet[T]) Remove(k T) bool {
 	m.mutex.Lock()
 	defer m.mutex.Unlock()
 
@@ -63,24 +63,24 @@ func (m *syncSet) Remove(k interface{}) bool {
 	return false
 }
 
-func (m *syncSet) Contains(k interface{}) bool {
+func (m *syncSet[T]) Contains(k T) bool {
 	m.mutex.Lock()
 	defer m.mutex.Unlock()
 	_, exists := m.m[k]
 	return exists
 }
 
-func (m *syncSet) Slice() []interface{} {
+func (m *syncSet[T]) Slice() []T {
 	m.mutex.Lock()
 	defer m.mutex.Unlock()
-	var a []interface{}
+	var a []T
 	for k, _ := range m.m {
 		a = append(a, k)
 	}
 	return a
 }
 
-func (m *syncSet) ForEach(f func(interface{})) {
+func (m *syncSet[T]) ForEach(f func(T)) {
 	m.mutex.Lock()
 	defer m.mutex.Unlock()
 	for k, _ := range m.m {
@@ -88,7 +88,7 @@ func (m *syncSet) ForEach(f func(interface{})) {
 	}
 }
 
-func (m *syncSet) ForEachFailFast(f func(interface{}) error) error {
+func (m *syncSet[T]) ForEachFailFast(f func(T) error) error {
 	m.mutex.Lock()
 	defer m.mutex.Unlock()
 	for k, _ := range m.m {
@@ -100,18 +100,18 @@ func (m *syncSet) ForEachFailFast(f func(interface{}) error) error {
 	return nil
 }
 
-func (m *syncSet) ForEachAsync(f func(interface{})) {
+func (m *syncSet[T]) ForEachAsync(f func(T)) {
 	for _, k := range m.Slice() {
 		f(k)
 	}
 }
 
-func (m *syncSet) Iterator() Iterator {
+func (m *syncSet[T]) Iterator() Iterator[T] {
 	a := m.Slice()
-	return NewIterator(a...)
+	return NewIterator[T](a...)
 }
 
-func (m *syncSet) ReverseIterator() Iterator {
+func (m *syncSet[T]) ReverseIterator() Iterator[T] {
 	a := m.Slice()
 	a = reverseSlice(a)
 	return NewIterator(a...)
